Move PaintRegion to separate file
